requirements.txt
rumps==0.4.0
py2app==0.28.10
Pillow==12.2.0